Comment up new utils and rename one
